Implement head removal from queue
Implement q_remove_head to remove an element from the queue head.
The function first checks whether the queue is NULL or empty.
If valid, it retrieves the first element. When a string buffer
is provided, it copies the element's value into the buffer using
strncpy and ensures null termination. Finally, the element is
removed from the list and its pointer is returned.

Change-Id: I1d3801e67aac2be0698f311a1741a3cbf5340b18