Reverse elements in queue
Implement q_reverse to reverse the order of elements in a queue.
The function checks if the head pointer is NULL or if the queue is
empty before proceeding. It then iterates over each node using
list_for_each_safe and moves each node to the head using list_move.
This effectively reverses the order of nodes in the queue.

Change-Id: I07fc72b7fff7727f5dd115e3b9252ee0a60e7f73