Remove nodes with higher value on right
Implement q_descend to remove each node with a higher value.
It removes nodes that have a higher value in any node to their right.
It checks if the queue is empty or singular before iterating.
For each node, it scans the right for a higher value.
When found, the node is removed and its memory freed.
Finally, the updated queue size is returned.

Change-Id: I439062425946f9aa90848313a8e14605c451c710