Sort queue elements in order
Implement q_sort to recursively sort the elements of a queue in either
ascending or descending order. The function locates the midpoint of the
queue, divides it into two halves, and then recursively sorts each half.
After sorting, it merges the two halves using q_merge_two with the
descend flag to determine the sort order. This commit adds a flexible
sorting capability to the queue.

Change-Id: I31ca2ea264de383866a7ea1e2555d30a857f13f9